pyproject.toml
[build-system]
requires = ["setuptools>=61.2.0", "wheel", "setuptools_scm[toml]>=3.4.3"]
build-backend = "setuptools.build_meta"

[project]
name = "JavaImageTools"
readme = "ReadMe.md"
description = "A library bridging PIL (Pillow) to `java.awt.image` and `javax.imageio`"
authors = [{name = "KOLANICH"}]
license = {text = "Unlicense"}
classifiers = [
	"Development Status :: 4 - Beta",
	"Environment :: Other Environment",
	"Intended Audience :: Developers",
	"License :: Public Domain",
	"Operating System :: OS Independent",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
	"Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["JavaImageTools"]
requires-python = ">=3.4"
dependencies = [
	"JAbs", # @ git+https://codeberg.org/KOLANICH-libs/JAbs.py
	"JPype1", # for now, not yet wrapped into abstraction layer
	"numpy",
	"Pillow",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://codeberg.org/KOLANICH-libs/JavaImageTools.py"

[tool.setuptools]
zip-safe = true
include-package-data = false

[tool.setuptools.packages]
find = {namespaces = false}

[tool.setuptools_scm]
